+++ src/App.jsx
@@ -1,20 +1,23 @@
+import {
+  BrowserRouter as Router, Routes,
+  Route,
+} from 'react-router-dom';
+import Navbar from './components/Navbar.jsx';
+import Books from './components/Books.jsx';
+import Category from './components/Categories.jsx';
+import Form from './components/Form.jsx';
+
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div>
+      <Router>
+        <Navbar />
+        <Routes>
+          <Route path="/" element={<Books />} />
+          <Route path="/category" element={<Category />} />
+          <Route path="/form" element={<Form />} />
+        </Routes>
+      </Router>
     </div>
   );
 }
