--- src/components/form.jsx
@@ -1,9 +1,0 @@
-const Form = () => (
-    <form className="form">
-      <input type="text" name="author" placeholder="Author" className="author" />
-      <input type="text" name="book" placeholder="Book" className="book" />
-      <input type="submit" value="Submit" className="submit" />
-    </form>
-  );
-  
-  export default Form;
