+++ src/index.css
@@ -1,13 +1,96 @@
 body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+  background-color: rgb(250, 250, 250);
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
+* {
+  padding: 0;
+  margin: 0;
+  text-decoration: none;
+}
+
+.navbar {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  background-color: #fff;
+  padding-left: 3rem;
+  height: 50px;
+}
+
+.links {
+  display: flex;
+  flex-direction: row;
+  gap: 2rem;
+  margin-left: 3rem;
+}
+
+.links a {
+  color: black;
+  text-decoration: none;
+}
+
+.card {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 1.5rem;
+  background-color: #fff;
+  padding: 2rem;
+  align-items: center;
+}
+
+.card ul {
+  list-style: none;
+}
+
+button {
+  background-color: rgb(2, 144, 255);
+  color: #fff;
+  border: none;
+  padding: 6px;
+  border-radius: 2px;
+}
+
+.form {
+  display: flex;
+  flex-direction: row;
+  margin: 2rem;
+  gap: 2rem;
+  justify-content: center;
+}
+
+.author {
+  width: 180px;
+}
+
+.book {
+  width: 550px;
+}
+
+.submit {
+  width: 100px;
+}
+
+input {
+  padding: 0.4rem;
+}
+
+span {
+  opacity: 0.5;
+  font-weight: bold;
+  font-size: medium;
+}
+
+p {
+  color: rgb(67, 134, 191);
+}
+
+.action-buttons {
+  display: flex;
+  flex-direction: row;
+  gap: 1px;
+}
+
+label {
+  margin-left: 50px;
 }
